Added fixes to support 2015r2, Keychain services, and various changes
Corrected several errors which prevented the classes from compiling
with 2015r2 due to incorrect lib names.  Also added KeychainServices
and an example view to demonstrate their use, and corrected the Email
example to follow Apple recommendations and recreate the
MFMailComposeViewController for each email.  Corrected the
LoadFramework method to return the handle of the library on first load
(a oversight prevented this) and added an empty StoreKit module to
contain the StoreKit classes when they are finished.